Back/requirements.txt
albumentations~=1.0.3
matplotlib~=3.5.1
torchvision~=0.1.6
numpy~=1.21.6
scipy~=1.4.1
tqdm~=4.64.0
scikit-learn~=1.0.2
tensorflow~=1.14.0
pillow~=9.0.1